.github/actions/publish-to-confluence/requirements.txt
atlassian-python-api
markdown
